# api-contracts/unified-openapi.yaml
openapi: 3.1.0
info:
  title: MyTrader Unified API
  version: 1.1.0
  description: |
    ## MyTrader Trading Platform Unified API Specification

    **CONTRACT VERSION**: 1.1.0
    **EFFECTIVE DATE**: 2024-09-24
    **PRIORITY IMPLEMENTATION**: Phase 1 - WebSocket Standardization, Phase 2 - API Versioning

    ### Critical Issues Addressed:
    1. **Port Configuration**: Standardized on port 5002 (backend) for all environments
    2. **WebSocket Event Naming**: Unified event naming without "Receive" prefixes
    3. **API Versioning**: Consistent `/api/v1/` pattern with backward compatibility
    4. **Volume Leaders Endpoint**: New endpoint for top volume assets per class

    ### Access Patterns
    - **Public Access**: Market data, dashboard, real-time updates (no authentication required)
    - **Authenticated Access**: Trading operations, user profiles, personal data (JWT Bearer token required)

    ### Supported Asset Classes
    - **CRYPTO** - 24/7 trading via multiple data providers
    - **STOCK_BIST** - Borsa Istanbul during Turkish market hours
    - **STOCK_NASDAQ** - US stocks during market hours
    - **FOREX** - Currency pairs (future implementation)

    ### Port Configuration (CRITICAL FIX)
    - **Backend API**: Port 5002 (all environments)
    - **WebSocket Hubs**: Port 5002 (unified with API server)
    - **Frontend Web**: Connects to port 5002 (not 8080)
    - **Frontend Mobile**: Connects to port 5002

...

...

  contact:
...
servers:
...
security:
  - bearerAuth: []
  - {}  # Allow anonymous access for public endpoints
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: Create a new user account with email verification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and return JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSessionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh JWT token
      description: Exchange refresh token for new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get user profile
      description: Get current user's profile information
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags: [Authentication]
      summary: Update user profile
      description: Update current user's profile information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateProfileResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: User logout
      description: Invalidate current session and revoke tokens
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ====================================================================
  # SYMBOLS MANAGEMENT (v1)
  # ====================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token obtained from login endpoint
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    UnauthorizedError:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Resource already exists or conflict occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          minLength: 8
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]'
          description: Password with minimum complexity requirements
        firstName:
          type: string
          minLength: 1
          maxLength: 50
        lastName:
          type: string
          minLength: 1
          maxLength: 50
        acceptTerms:
          type: boolean
          description: Must be true to register
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        rememberMe:
          type: boolean
          default: false
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 1
    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 50
        lastName:
          type: string
          minLength: 1
          maxLength: 50
        email:
          type: string
          format: email
    RegisterResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        userId:
          type: string
          format: uuid
        verificationRequired:
          type: boolean
    UserSessionResponse:
      type: object
      required: [success, user, accessToken, refreshToken, tokenExpiry]
      properties:
        success:
          type: boolean
        user:
          $ref: '#/components/schemas/UserProfile'
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenExpiry:
          type: string
          format: date-time
        sessionId:
          type: string
          format: uuid
    UserProfile:
      type: object
      required: [id, email, firstName, lastName, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        isEmailVerified:
          type: boolean
        avatar:
          type: string
          format: uri
        preferences:
          type: object
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
    UserProfileResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserProfile'
    UpdateProfileResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserProfile'
    TokenResponse:
      type: object
      required: [accessToken, refreshToken, tokenExpiry]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenExpiry:
          type: string
          format: date-time
    LogoutResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'

    # ================================================================
    # SYMBOL MANAGEMENT SCHEMAS
    # ================================================================
...
    BaseResponse:
      type: object
      required: [success, timestamp]
      properties:
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              type: boolean
              enum: [false]
            error:
              type: object
              required: [code, message]
              properties:
                code:
                  type: string
                message:
                  type: string
                details:
                  type: string
                field:
                  type: string
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              type: boolean
              enum: [false]
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
...